修改过滤条件bug

+++ src/models/calendar/index.js
@@ -204,8 +204,7 @@ export default {
           // item.flag == '0'
           // duetimeStamp = duetimeStamp && duetimeStamp.length < 13 ? duetimeStamp * 1000 : duetimeStamp;
 
-
-          if (parseInt(item.time_warning) > 0 && item.flag == '0' && item.is_realize == '0' && !(duetimeStamp < timeStamp)) {
+          if (parseFloat(item.time_warning) > 0 && item.flag == '0' && item.is_realize == '0' && !(duetimeStamp < timeStamp)) {
             var item1 = {
               time: due_time - 86400000 * parseInt(item.time_warning),
               type: 2,
